Place repair task on the same Cassandra node that needs to be repaired
Summary:
Resolves T506629. We hit an issue that repair job was placed on a node other than the one needs to be repaired.
The consequent problem is that the executor only connects to the Cassandra daemon on the local node. So it may
not work and keep trying. The same issue exists for other maintenance task like backup/restore.

To fix the issue, this diff introduces a new placement strategy and is supposed to be used by all maintenance
tasks that need to be placed on the same Cassandra node.

Be noted that the latest framework (1.0.14+) uses a different approach but it involves more changes that we're
not ready to opt in at this moment.

Test Plan:
Tested locally with 5 nodes that all (10+) repair tasks were placed properly.
Also tested with staging-sjc1 (6 nodes plus 3 reserved nodes) that all of 120 repair tasks were placed properly.

Reviewers: zjay, karthik, zhiyan, varung, verma

Reviewed By: verma

Maniphest Tasks: T506629

Revert Plan: NA

Differential Revision: https://code.uberinternal.com/D548204